do_email/requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
flask>=2.0.0
pytz>=2021.1
mysql-connector-python>=8.0.0
